+++ packages/tools/src/_utils/dom.ts
@@ -1,5 +1,7 @@
 import qq from '@hugsmidjan/qj/qq';
 
+import { HTMLText, Regulation } from '../types';
+
 export const styleValueToTypeAttrMap: Record<string, string | undefined> = {
   circle: 'circle',
   square: 'square',
@@ -28,7 +30,31 @@ export const asDiv = (html: string) => {
 };
 
 // ---------------------------------------------------------------------------
+
 export const document_base_url =
   typeof document !== 'undefined'
     ? document.location.href.replace(/^(https?:\/\/[^/]+).+/i, '$1')
     : '';
+
+// ---------------------------------------------------------------------------
+
+export const combineTextAppendixesComments = (
+  text: HTMLText,
+  appendixes: Regulation['appendixes'] | undefined,
+  comments: HTMLText | undefined,
+): HTMLText => {
+  const wrappedAppendixes = (appendixes || [])
+    .map(
+      ({ title, text }) =>
+        `<section class="appendix">` +
+        `  <h2 class="appendix__title">${title.replace(/</g, '&lt;')}</h2>` +
+        `  ${text}` +
+        `</section>`,
+    )
+    .join('');
+
+  const wrappedComments =
+    comments && `<section class="comments">${comments}</section>`;
+
+  return (text + wrappedAppendixes + wrappedComments) as HTMLText;
+};